Crash fix in chat.cpp
How to reproduce a crash?
Just login, and send messages. Navigate into history (Shift + Up or Down) and this will work normally. But if you relog your character and try to navigate (Shift + Up) the client will crash.

We need to check if "m_savedMessages" is empty, because if this is empty we can't navigate in a empty history.